requirements.txt
mysql
streamlit
open-python
pillow
pyzbar
pymysql
os
mysql-connector-python
pandas